+++ src/Physics/Debug.h
@@ -22,20 +22,20 @@ bool assertFailedCallback(const char *expression, const char *message, const cha
 #ifdef JPH_DEBUG_RENDERER
 class BatchImpl : public JPH::RefTargetVirtual, public JPH::RefTarget<BatchImpl> {
    private:
-    size_t index_ = 0;
-    size_t vertex_ = 0;
-    size_t count_ = 0;
+    uint32_t index_ = 0;
+    uint32_t vertex_ = 0;
+    uint32_t count_ = 0;
 
    public:
-    BatchImpl(size_t index, size_t vertex, size_t count)
+    BatchImpl(uint32_t index, uint32_t vertex, uint32_t count)
         : index_(index), vertex_(vertex), count_(count) {
     }
 
-    size_t count() const {
+    uint32_t count() const {
         return count_;
     }
 
-    size_t baseVertex() const {
+    uint32_t baseVertex() const {
         return vertex_;
     }
